backend/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Data Validation & ORM
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlmodel>=0.0.14

# Database
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
argon2-cffi>=23.1.0

# Email validation
email-validator>=2.1.0

# Observability
structlog>=24.1.0
prometheus-client>=0.19.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0

# Database Migrations
alembic>=1.13.0

# AI Integration (Phase 3)
openai>=1.12.0
anthropic>=0.18.0

# Phase 5: Advanced Features
apscheduler>=3.10.0
python-dateutil>=2.8.2
dapr>=1.12.0
